Move timing to util
