+++ libopenage/crossplatform/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_sources(libopenage
 	subprocess.cpp
 	os.cpp
-	timing.cpp
 )
