+++ libopenage/rng/rng.cpp
@@ -6,7 +6,7 @@
 #include <sstream>
 #include <string.h>
 
-#include "../crossplatform/timing.h"
+#include "../util/timing.h"
 #include "../error/error.h"
