+++ libopenage/util/fps.cpp
@@ -1,10 +1,10 @@
-// Copyright 2013-2015 the openage authors. See copying.md for legal info.
+// Copyright 2013-2016 the openage authors. See copying.md for legal info.
 
 #include "fps.h"
 
 #include <math.h>
 
-#include "../crossplatform/timing.h"
+#include "timing.h"
 
 namespace openage {
 namespace util {
