+++ libopenage/util/timer.cpp
@@ -1,8 +1,8 @@
-// Copyright 2013-2015 the openage authors. See copying.md for legal info.
+// Copyright 2013-2016 the openage authors. See copying.md for legal info.
 
 #include "timer.h"
 
-#include "../crossplatform/timing.h"
+#include "../util/timing.h"
 
 namespace openage {
 namespace util {
