+++ libopenage/util/timing.cpp
