+++ libopenage/util/timing.h
